Revert "Try MinDiskFree.value as Long instead of Integer"
This reverts commit d3588629818b3f75d30a2486a7c4583b0d4c99c7.

+++ app/src/main/java/com/nutomic/syncthingandroid/model/Folder.java
@@ -51,7 +51,7 @@ public class Folder {
     }
 
     public static class MinDiskFree {
-        public Long value = 1l;
+        public Integer value = 1;
         public String unit = "%";
     }
